stuck on server stuff

+++ src/main/RDFItem.java
@@ -25,7 +25,7 @@ public class RDFItem extends Item {
 		super();
 	}
 	
-	public RDFItem(String cust, String prodName, String unit, String gtin, Date packDate, int qty, float price, String itemCode) {
+	public RDFItem(String cust, String prodName, String unit, String gtin, Date packDate, float qty, float price, String itemCode) {
 		super(cust, prodName, unit, gtin, packDate, qty, price, itemCode);
 	}
